# restaurant-pos-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Tamanna Restaurant POS API
  description: >
    Comprehensive REST API documentation for Tamanna Restaurant Pure Vegetarian POS & ERP System.
    Includes Point of Sale, KDS, Dynamic Tables, Razorpay Online Gateway, SMS/OTP, Automated Cron Scheduling,
    Excel/PDF document exports, and Real-time WebSockets.
  version: 1.0.0
  contact:
    name: Tamanna Restaurant IT Support
    email: support@tamannarestaurant.com
servers:
  - url: http://localhost:5000
    description: Local Development & POS Server
paths:
...
  /api/tables:
    get:
      summary: List all restaurant tables with live occupancy status
      tags:
        - Tables
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Array of tables
  /api/orders:
    get:
      summary: Fetch active and historical dining orders
      tags:
        - Orders
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of orders
    post:
      summary: Create new order / KOT
      tags:
        - Orders
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tableId, items]
              properties:
                tableId: { type: string }
                items:
                  type: array
                  items:
                    type: object
                    properties:
                      name: { type: string }
                      price: { type: number }
                      quantity: { type: integer }
      responses:
        '201':
          description: Order created and KOT broadcasted
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT